refactor(workspace): collapse exec helpers, runtime, and ExecOneShot into exec.go
Four files (exec_helpers.go, runtime.go, docker_runtime.go, exec_oneshot.go)
covered one logical surface: container exec via a runtime abstraction.
Merge them into a single exec.go organized by responsibility:

  1. workspace metadata helpers (ResolveDockerCommand, LoadExecResult, ...)
  2. ContainerRuntime interface + ExecRequest/ContainerTarget types
  3. DockerRuntime implementation
  4. ExecOneShot + ResolveTimeout high-level entry point

Rename exec_oneshot_test.go to exec_test.go to match.
